fixes and tests

+++ app/Services/Contact/Relationship/DestroyRelationship.php
@@ -36,14 +36,16 @@ class DestroyRelationship extends BaseService
 
         $reverseRelationshipType = $relationship->account->getRelationshipTypeByType($relationship->relationshipType->name_reverse_relationship);
 
-        $reverseRelationship = Relationship::where('account_id', $data['account_id'])
-                                            ->where('contact_is', $relationship->of_contact)
-                                            ->where('of_contact', $relationship->contact_is)
-                                            ->where('relationship_type_id', $reverseRelationshipType->id)
-                                            ->first();
+        if ($reverseRelationshipType) {
+            $reverseRelationship = Relationship::where('account_id', $data['account_id'])
+                                                ->where('contact_is', $relationship->of_contact)
+                                                ->where('of_contact', $relationship->contact_is)
+                                                ->where('relationship_type_id', $reverseRelationshipType->id)
+                                                ->first();
 
-        if ($reverseRelationship) {
-            $reverseRelationship->delete();
+            if ($reverseRelationship) {
+                $reverseRelationship->delete();
+            }
         }
 
         $relationship->delete();
